refactor navBar
