+++ src/widgets/navBar/components/navBar/navBar.tsx
@@ -1,5 +1,5 @@
 import MenuList from '../menuList/menuList';
-import { navBar } from '../../api/menuItems';
+import { navBar } from '../../model/menuItems';
 import { IMenuItems } from '../../model/menuItemsTypes';
 
 import './navBar.scss'
