+++ src/widgets/navBar/model/menuItems.ts
